Add Couchbase support to the web service. Add option to skip S3 upload for testing purposes.

+++ camera-webservice/webservice.py
@@ -4,7 +4,11 @@ from flask import Flask, jsonify, abort, make_response, request
 from flask_restful import Api, Resource, reqparse, fields, marshal
 import time, socket, uuid, os, boto3, json, sys, threading
 from datetime import datetime
+import uuid
 from elasticsearch import Elasticsearch
+from couchbase.cluster import Cluster
+from couchbase.cluster import PasswordAuthenticator
+from couchbase.n1ql import N1QLQuery
 if os.name == 'nt':
 	from PIL import Image
 
@@ -12,6 +16,13 @@ if os.name == 'nt':
 import botocore.vendored.requests
 from botocore.vendored.requests.packages.urllib3.exceptions import InsecureRequestWarning
 botocore.vendored.requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
+
+def open_couchbase_bucket(cluster, bucket, username, password) :
+	cluster = Cluster(cluster)
+	authenticator = PasswordAuthenticator(username, password)
+	cluster.authenticate(authenticator)
+	bucket = cluster.open_bucket(bucket)
+	return bucket
 
 # Get own IP address
 def get_ip_address():
@@ -52,6 +63,17 @@ def heartbeat_es_index(elasticsearch_host, ip_address, camera_name):
 		else:
 			print ('## ElasticSearch heartbeat update failed')
 		time.sleep(60)
+
+		# Update camera index every 60s as a heartbeat to determine 'active' cameras
+def heartbeat_cb_bucket(ip_address, camera_name):
+	while (1):
+		ts = round(time.time())
+		key = 'camera::' + ip_address
+		camera_info = {'camera_name': conf['camera_name'], 'epoch_timestamp': ts, 'camera_ip': ip_address}
+		res = bucket.upsert(key, camera_info)
+		print ('## Couchbase heartbeat saved at: id=[' + key + '], epoch_timestamp=[' + str(ts) + ']')
+		time.sleep(60)
+
 
 class RootAPI(Resource):
 
@@ -96,16 +118,22 @@ class TakePhotoAPI(Resource):
 		except:
 			pass
 
-		# Upload to S3 endpoint
-		session = boto3.session.Session(aws_access_key_id=conf['access_key'], aws_secret_access_key=conf['secret_access_key'])
-		s3 = session.resource(service_name='s3', endpoint_url=conf['endpoint'], verify=False)
-		obj = s3.Object(conf['bucket'], filename)
-		data = open(filename, 'rb')
-		obj.put(Body=data, ContentType='image/jpeg')
-		data.close()
-		os.remove(filename)
-		url = conf['endpoint'] + "/" + conf['bucket'] + "/" + filename
-		print ('## Photo saved at: ' + url)
+		# For testing to skip S3 posting
+		#  curl -s -H "Content-Type: application/json" -X POST -d '{"skip_s3":true}' http://localhost:8080/take_photo
+
+		if 'skip_s3' in req and req['skip_s3'] is True:
+			url = "http://localhost/test.jpg"
+		else:
+			# Upload to S3 endpoint
+			session = boto3.session.Session(aws_access_key_id=conf['access_key'], aws_secret_access_key=conf['secret_access_key'])
+			s3 = session.resource(service_name='s3', endpoint_url=conf['endpoint'], verify=False)
+			obj = s3.Object(conf['bucket'], filename)
+			data = open(filename, 'rb')
+			obj.put(Body=data, ContentType='image/jpeg')
+			data.close()
+			os.remove(filename)
+			url = conf['endpoint'] + "/" + conf['bucket'] + "/" + filename
+			print ('## Photo saved at: ' + url)
 
 		# Post image taken to ES; let id be completed by ES
 		es = Elasticsearch([conf['elasticsearch_host']])
@@ -119,6 +147,10 @@ class TakePhotoAPI(Resource):
 		else:
 			print ('## ElasticSearch update failed')
 
+		key = str(uuid.uuid4())
+		res = bucket.upsert(key, res_body)
+		print ('## Couchbase saved at:  id=[' + key + ']')
+
 		# Return success
 		return make_response(jsonify(res_body), 200)
 
@@ -128,6 +160,8 @@ api = Api(app)
 
 api.add_resource(RootAPI, '/')
 api.add_resource(TakePhotoAPI, '/take_photo')
+
+global bucket
 
 if __name__ == '__main__':
 
@@ -142,15 +176,31 @@ if __name__ == '__main__':
 			conf['secret_access_key']
 			conf['elasticsearch_host']
 			conf['camera_command']
+			conf['couchbase_host']
+			conf['couchbase_bucket']
+			conf['couchbase_username']
+			conf['couchbase_password']
 	except Exception as e:
 		sys.stderr.write('FATAL: Cannot open or parse configuration file : ' + str(e) + '\n\n')
+		exit()
+
+	try:
+		global bucket
+		bucket = open_couchbase_bucket(conf['couchbase_host'], conf['couchbase_bucket'], conf['couchbase_username'], conf['couchbase_password'])
+	except Exception as e:
+		sys.stderr.write('FATAL: Cannot open Couchbase bucket : ' + str(e) + '\n\n')
 		exit()
 
 	ip_address = get_ip_address()
 	#ip_address = '1.2.3.4'
 	create_es_indicies()
-	t = threading.Thread(target=heartbeat_es_index, args=(conf['elasticsearch_host'], ip_address, conf['camera_name']))
-	t.daemon = True
-	t.start()
+	t_es = threading.Thread(target=heartbeat_es_index, args=(conf['elasticsearch_host'], ip_address, conf['camera_name']))
+	t_es.daemon = True
+	t_es.start()
+
+	t_cb = threading.Thread(target=heartbeat_cb_bucket, args=(ip_address, conf['camera_name']))
+	t_cb.daemon = True
+	t_cb.start()
+
 	print ('## camera-webservice will be reachable at http://'+ ip_address + ':8080')
 	app.run(host='0.0.0.0', port=8080, debug=True, use_reloader=False)
